--- app/src/main/java/com/example/finalproject/Intruction.java
@@ -1,14 +1,0 @@
-package com.example.finalproject;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class Intruction extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_intruction);
-    }
-}
