+++ app/src/main/res/values/strings.xml
@@ -39,10 +39,7 @@
     <string name="contact">contact</string>
     <string name="search">search</string>
     <string name="version">version: 2.0</string>
-    <string name="liao">Mingjiong Liao  liao0019@algonquinlive.com</string>
-    <string name="wang">Qing Wang       wang0582@algonquinlive.com</string>
-    <string name="ye">Rongzhong Ye    ye000038@algonquinlive.com</string>
-    <string name="luo">Chunyuan Luo    luo00047@algonquinlive.com</string>
+
     <string-array name="spinner1">
         <item>CAD</item>
         <item>USD</item>
